refactor(p4a): delete dead Puppeteer entrypoint, remove src/lab, move Observer to src/telemetry
Dead code removal:
- Delete qa/harness/installPuppeteerRuntime.ts — exported installLabRuntime
  but had zero callers anywhere in the codebase. Audit confirmed no external
  injector, no call site in src/, qa/, tests/, or scripts/.
- Delete src/lab/labRuntimeHost.ts — its only consumer was
  installPuppeteerRuntime.ts. LabRuntimeHost was the Puppeteer host interface
  and has no place in src/.
- Delete src/lab/debug/ports/index.ts and yaosDebugPort.ts — redundant shims
  pointing at src/telemetry/debug/ports/; zero importers.

Dead telemetry mount API removal:
- Remove onTelemetryApiMounted / onTelemetryApiUnmounted from TelemetryRuntimeHost.
  installTelemetryRuntime never called onTelemetryApiMounted (the mount was
  asymmetric — unmount was called, mount was not). Telemetry is passive and
  must not mount window.__YAOS_DEBUG__ — that is a Puppeteer/QA control surface,
  not a telemetry surface. The defensive delete of __YAOS_DEBUG__ in onunload
  remains sufficient.
- Remove the corresponding dead callbacks from the host object literal in main.ts.
- Remove the host.onTelemetryApiUnmounted() call from installTelemetryRuntime
  dispose() — nothing to unmount since nothing was ever mounted.

src/lab → src/telemetry rename:
- Move all passive Observer files out of the lying src/lab/ directory:
    src/lab/debug/{flightEvents,flightRecorder,flightTraceController,
                   flightTraceSink,pathIdentity,trace}.ts
    src/lab/diagnostics/{deviceWitnessTracker,diagnosticsBundle,
                         diagnosticsService,pathRedactor,witnessStateHash}.ts
  → src/telemetry/debug/ and src/telemetry/diagnostics/ respectively.
- Update all import references in src/, qa/, tests/ (53 files total).
- src/lab/ is now completely gone.

Guard updates:
- Remove installLabRuntime pattern from guard-qa-isolation.mjs TELEMETRY_FORBIDDEN
  (the file is deleted).

Verification:
  npm run build                       PASS
  npm run build:qa-product            PASS
  npm run guard:production-bundles:strict  PASS
  npm run guard:qa-isolation          PASS
  npm run verify:bundles              PASS
  npm run test:regressions            84/84
  npm run lint:changed                PASS
  rg 'lab/' src/ qa/ tests/ scripts/ → 0 (code files)
  rg 'LabRuntimeHost|installLabRuntime|onTelemetryApiMounted|onTelemetryApiUnmounted' → 0
