--- src/lab/debug/ports/yaosDebugPort.ts
@@ -1,5 +1,0 @@
-/**
- * YaosDebugPort — re-exported from the canonical telemetry location.
- * Canonical: src/telemetry/debug/ports/yaosDebugPort.ts
- */
-export type { YaosDebugPort, EditorBindingHealth, ReceiptSnapshot } from "../../../telemetry/debug/ports/yaosDebugPort";
