+++ src/telemetry/diagnostics/deviceWitnessTracker.ts
