+++ src/telemetry/diagnostics/diagnosticsBundle.ts
