+++ src/telemetry/diagnostics/diagnosticsService.ts
