+++ src/telemetry/diagnostics/pathRedactor.ts
