+++ src/telemetry/diagnostics/witnessStateHash.ts
