+++ tests/device-witness-qa-api.ts
@@ -9,9 +9,9 @@
  */
 
 import assert from "node:assert/strict";
-import { DeviceWitnessTracker } from "../src/lab/diagnostics/deviceWitnessTracker.js";
-import type { WitnessTrackerConfig } from "../src/lab/diagnostics/deviceWitnessTracker.js";
-import type { FlightSink, TraceContext } from "../src/lab/debug/flightEvents.js";
+import { DeviceWitnessTracker } from "../src/telemetry/diagnostics/deviceWitnessTracker.js";
+import type { WitnessTrackerConfig } from "../src/telemetry/diagnostics/deviceWitnessTracker.js";
+import type { FlightSink, TraceContext } from "../src/telemetry/debug/flightEvents.js";
 
 // -----------------------------------------------------------------------
 // Harness
