+++ tests/witness-readonly-spy.ts
@@ -11,8 +11,8 @@
  */
 
 import assert from "node:assert/strict";
-import { DeviceWitnessTracker } from "../src/lab/diagnostics/deviceWitnessTracker";
-import type { WitnessTrackerConfig } from "../src/lab/diagnostics/deviceWitnessTracker";
+import { DeviceWitnessTracker } from "../src/telemetry/diagnostics/deviceWitnessTracker";
+import type { WitnessTrackerConfig } from "../src/telemetry/diagnostics/deviceWitnessTracker";
 
 let passed = 0;
 let failed = 0;
